feat: Adding further type definitions; This will serve as a base for the type-safe modular approach we will later take. Created a types submodule that will later serve as a base for the whole FRSS library type system. Added support for creating multiple context pad entries per Element module.
Problem: the user interface for creating the control (pad / palette) entries is too modular
and too unintuitive (long chains of input arguments, very janky atm).

Solution: Extend and rewrite the portion of the type system to allow for a more user-friendly
approach. All parameters will be grouped into "categories" and will serve as the modular settings
for creating necessary entries for the bpmn-js in a user-friendly manner.

Think Prisma's modular type system.

Future problem: The types heavily rely on using `any` as the interface with the JS code, as well
as allowing keys to be undefined if necessary. This is HEAVILY discouraged and I will work on
eliminating most, if not all of the occurrences.

The key to solve this problem is module definitions for the code that interfaces with
the JS implementation of bpmn-js. Whether to go down on this route is up for the debate.