docs: Need to specify what object triggers rendering in rendererEntry.ts
This will likely result in re-thinking the type system. Problem arose when I tried to render the changed dataObjectRef. Need to add the identifier for the element to run the rendering job at, does not always make sense to run it on the identifier of that element.
Also should add a way to add conditions for pad entries -> you want to be able to unmark the DataObject as evidence -> there should be a way to conditionally display the entries ("mark as" on non-evidence, "unmark as" on evidence...)

Potential for type should be `(any) => bool | "always" | "never"`. The arrow function defines the condition and probably takes the element to scan it first